requirements.txt
fastapi
uvicorn
langchain
openai
chromadb
python-docx
pdfminer.six
tiktoken
pydantic
python-multipart
